requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=9.0.0
numpy>=1.23.0
ipywidgets>=8.0.0
ipython>=8.0.0
matplotlib>=3.6.0
